+++ trunk/xorp/bgp/open_packet.cc
@@ -12,7 +12,7 @@
 // notice is a summary of the XORP LICENSE file; the license in that file is
 // legally binding.
 
-#ident "$XORP: xorp/bgp/open_packet.cc,v 1.8 2003/01/29 20:32:31 rizzo Exp $"
+#ident "$XORP: xorp/bgp/open_packet.cc,v 1.9 2003/01/29 23:38:12 rizzo Exp $"
 
 #include "bgp_module.h"
 #include "config.h"
@@ -98,7 +98,7 @@ OpenPacket::OpenPacket(const uint8_t *d, uint16_t l)
 
     while (i > 0) {
 	size_t len;
-	debug_msg("Length of unread parameters : %d\n", i);
+	debug_msg("Length of unread parameters : %u\n", (uint32_t)i);
 
 	BGPParameter *p = BGPParameter::create(d, i, len);
 	if (p != NULL)
